jetson-inference-app/requirements.txt
torch==1.10.0+cu102
torchvision==0.11.1+cu102
numpy==1.21.2
Pillow==8.3.1
opencv-python==4.5.3.20210929
tqdm==4.62.3
matplotlib==3.4.3